moved slides data to slides store
